move test in tests/
